+++ web/app/audio/player.ts
@@ -119,3 +119,7 @@ export function initializeFromGesture() {
         createNewContext();
     }
 }
+
+export function globalAudioContext() : AudioContext {
+    return context();
+}