+++ ChangeLog.md
@@ -1,10 +1,14 @@
 # Changelog:
-** 22.11.20**
+* **29.11.20**
+    - Added support for the native client to show and broadcast video
+    - By default using a quick select method when sharing the screen
+    
+* **22.11.20**
     - Added a ton of video settings
     - Added screen sharing (Currently via the camera channel)
     - Using codec H264 instead of VP8
     
-** 14.11.20**
+* **14.11.20**
     - Fixed bug where the microphone has been requested when muting it.
     
 * **07.11.20**